+++ modules/middleware/logger.js
@@ -1,5 +1,4 @@
 var strftime = require('strftime').strftime;
-var microtime = require('microtime');
 
 /**
  * A middleware that writes log entry data about the response to a given stream.
@@ -10,24 +9,26 @@ function logger(app, outputStream) {
   outputStream = outputStream || process.stderr;
 
   return function (request) {
-    var start = microtime.now();
+    var start = Date.now();
 
     return request.call(app).then(function (response) {
-      var host = request.remoteHost || '-';
-      var id = '-'; // RFC 1413 identity of the client determined by identd on the client's machine
-      var user = request.remoteUser || '-';
-      var timestamp = '[' + strftime('%d/%b/%Y %H:%M:%S', request.date) + ']';
-      var info = '"' + request.method + ' ' + request.fullPath + ' HTTP/' + request.protocolVersion + '"';
-
+      var elapsedTime = (Date.now() - start) / 1000;
       var contentLength = response.headers['Content-Length'];
+
       if (contentLength == null)
         contentLength = '-';
 
-      var elapsedMicroseconds = microtime.now() - start;
-      var seconds = Math.round(elapsedMicroseconds / 100) / 10000;
-
       // 127.0.0.1 - frank [10/Oct/2000 13:55:36] "GET /apache_pb.gif HTTP/1.0" 200 2326 0.003
-      var entry = [ host, id, user, timestamp, info, response.status, contentLength, seconds ].join(' ');
+      var entry = [
+        request.remoteHost || '-',
+        '-', // RFC 1413 identity of the client
+        request.remoteUser || '-',
+        '[' + strftime('%d/%b/%Y %H:%M:%S', request.date) + ']',
+        '"' + request.method + ' ' + request.fullPath + ' HTTP/' + request.protocolVersion + '"',
+        response.status,
+        contentLength,
+        elapsedTime
+      ].join(' ');
 
       outputStream.write(entry + '\n');